+++ kerastuner/tuners/bayesian.py
@@ -5,11 +5,11 @@ import numpy as np
 from scipy import optimize as scipy_optimize
 from sklearn import gaussian_process
 
-from ..abstractions.tensorflow import TENSORFLOW_UTILS as tf_utils
 from ..engine import hyperparameters as hp_module
+from ..engine import multi_execution_tuner
 from ..engine import oracle as oracle_module
 from ..engine import trial as trial_lib
-from ..tuners import multi_execution_tuner
+from ..abstractions.tensorflow import TENSORFLOW_UTILS as tf_utils
 
 
 class BayesianOptimizationOracle(oracle_module.Oracle):
