+++ kerastuner/tuners/hyperband.py
@@ -15,10 +15,11 @@ import copy
 import queue
 import random
 import json
-from ..engine import trial as trial_lib
-from ..engine import oracle as oracle_module
+
 from ..abstractions.tensorflow import TENSORFLOW_UTILS as tf_utils
-from ..tuners import multi_execution_tuner
+from ..engine import multi_execution_tuner
+from ..engine import oracle as oracle_module
+from ..engine import trial as trial_lib
 
 
 def queue_to_list(queue):
