+++ kerastuner/tuners/randomsearch.py
@@ -18,14 +18,14 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from ..engine import trial as trial_lib
-from ..engine import oracle as oracle_module
-from ..engine import hyperparameters as hp_module
 from ..abstractions.tensorflow import TENSORFLOW_UTILS as tf_utils
-from ..tuners import multi_execution_tuner
+from ..engine import hyperparameters as hp_module
+from ..engine import multi_execution_tuner
+from ..engine import oracle as oracle_module
+from ..engine import trial as trial_lib
 
-import random
 import json
+import random
 
 
 class RandomSearchOracle(oracle_module.Oracle):
